Configure Vagrant to use 2 VMs:
We want to setup a test Puppet Master and agent(s) system.  We define
2 VMs, one for the master and the other as an agent.  These 2 VMs are
provisioned using puppet, it first checks if puppet is already installed
if so, it removes puppet and then installs puppet from the Puppet Labs
CentOS repo, this ensures there is a known base-level.  Once this is done
the machine is provisioned as normal, where Vagrant specifies the location
of puppet configuration files/manifests etc.