Added the Puppet master/agent nodes to be provisioned:
We define 2 nodes, one represents the Puppet Master, the other the Puppet Agent connecting
to the Master.  Very little is done here, we ensure Puppet is installed on both systems and
that Puppet Master is installed on the master.  We still have to configure the master to
accept requests from the client.  This should NOT be done automatically, this should just
provision the machine, we should then connect to the master manually so certificates are
correctly setup.